% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fevol}
\alias{fevol}
\title{Detects fast evolving sites}
\usage{
fevol(file, cutoff = 0.9)
}

\arguments{
    \item{file}{Sequence alignment file name specified by either a variable of mode character,
    or a double-quoted string; if \code{file = ""} (the default) then the
    program terminates with message}
    \item{cutoff}{OV value cutoff for fast evolving site detection. cutoff = 0.9 is set as default.}
}
\description{
To calculate observed variability (OV), all sequences for a given position are compared in a pair-wise fashion. Mismatches are scored as 1 and matches as 0; the mean value amongst all the comparisons for a given position is used as the measure of character variability in the subsequent data sorting. Check reference for details.
}

\references{
  Goremykin VV1, Nikiforova SV, Bininda-Emonds OR. 2010. Automated removal of noisy data in phylogenomic analyses. J Mol Evol. 71:319-31. 
  \url{http://www.ncbi.nlm.nih.gov/pubmed/20976444}

}

