% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rycode}
\alias{rycode}
\title{"RY-coding"}
\usage{
rycode(file, type = 3)
}

\arguments{
    \item{file}{Sequence alignment file name specified by either a variable of mode character,
    or a double-quoted string; if \code{file = ""} (the default) then the
    program terminates with message}
    \item{type}{RY coding position. If type=3 (Default=3), program performs RY coding at the third
    codon position whereas if type="all" then program performs RY coding for all the
    positions}
}
\description{
Class \code{"RY-coding"}
}

\author{Ambuj Kumar \email{ambuj@ufl.edu}}