% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.nexus}
\alias{write.nexus}
\title{Class \code{"write-nexus"}}
\usage{
write.nexus(x, file, format = "dna", datablock = TRUE, interleaved = TRUE,
  charsperline = NULL, gap = NULL, missing = NULL)
}
\description{
Class \code{"write-nexus"}
}

